refactor(runtime): descend resume and access primitives
Move the pure resume identity recipe into nika-proof and route receipts
into nika-providers. Keep runtime re-exports source-compatible while
restoring the crate-size ratchet without weakening its limit.
