+++ crates/nika-proof/Cargo.toml
@@ -12,8 +12,10 @@ publish                = false  # Foundation crate — never on crates.io. See A
 
 [dependencies]
 nika-check  = { path = "../nika-check", version = "0.113.0" }    # RunCertificate — the check certificate the receipt folds
+nika-schema = { path = "../nika-schema", version = "0.113.0" }   # raw task shapes in the resume identity recipe
 serde_json  = { workspace = true }
 blake3      = { workspace = true }   # the engine-side digest (the reference's sha256 · parity on the pre-image, never the digest)
+serde_json_canonicalizer = { workspace = true } # RFC 8785 bytes for resume keys
 
 [dev-dependencies]
 serde_json = { workspace = true }
